Fixed bug where unable to get the abs path if the file doesn't exist

+++ Build.ps1
@@ -1,4 +1,6 @@
 $ErrorActionPreference = 'Stop'
-$distFolder = resolve-path 'dist\'
-Remove-Item $distFolder -Force -Recurse
+$distFolder = $ExecutionContext.SessionState.Path.GetUnresolvedProviderPathFromPSPath('./dist/')
+if(Test-Path $distFolder){
+    Remove-Item $distFolder -Force -Recurse 
+}
 msbuild '/t:restore;build;publish' DBRestorer.sln /p:Configuration=Release /p:PublishDir=$distFolder /m
